Move some utils out of cli/core.py
