+++ indico/cli/core.py
@@ -16,15 +16,17 @@
 
 from __future__ import unicode_literals
 
-import traceback
-from importlib import import_module
-
 import click
-from flask.cli import FlaskGroup, pass_script_info, ScriptInfo, AppGroup
-from flask_pluginengine import wrap_in_plugin_context
-from werkzeug.utils import cached_property
+from flask.cli import pass_script_info, AppGroup
+
+# XXX: Do not import any other indico modules here!
+# If any import from this module triggers an exception the dev server
+# will die while an exception only happening during app creation will
+# be handled gracefully.
+from indico.cli.util import IndicoFlaskGroup, LazyGroup
 
 click.disable_unicode_literals_warning = True
+__all__ = ('cli_command', 'cli_group')
 
 
 # We never use the group but expose cli_command and cli_group for
@@ -34,91 +36,6 @@ _cli = AppGroup()
 cli_command = _cli.command
 cli_group = _cli.group
 del _cli
-__all__ = ('cli_command', 'cli_group')
-
-
-def _create_app(info):
-    # XXX: Keep this import local - importing this file should not
-    # import any other indico code as having the import of this file
-    # fail would break the reloader.
-    from indico.web.flask.app import make_app
-    return make_app(set_path=True)
-
-
-class IndicoFlaskGroup(FlaskGroup):
-    def __init__(self, **extra):
-        super(IndicoFlaskGroup, self).__init__(create_app=_create_app, add_default_commands=False,
-                                               add_version_option=False, **extra)
-        self._indico_plugin_commands = None
-
-    def _load_plugin_commands(self):
-        # We don't care about `flask.commands` but indico plugin commands instead
-        # This actually shouldn't be called sinde we override all the methods
-        # calling it...
-        assert False
-
-    def _wrap_in_plugin_context(self, plugin, cmd):
-        cmd.callback = wrap_in_plugin_context(plugin, cmd.callback)
-        for subcmd in getattr(cmd, 'commands', {}).viewvalues():
-            self._wrap_in_plugin_context(plugin, subcmd)
-
-    def _get_indico_plugin_commands(self, ctx):
-        if self._indico_plugin_commands is not None:
-            return self._indico_plugin_commands
-        try:
-            from indico.core import signals
-            from indico.util.signals import named_objects_from_signal
-            ctx.ensure_object(ScriptInfo).load_app()
-            cmds = named_objects_from_signal(signals.plugin.cli.send(), plugin_attr='_indico_plugin')
-            rv = {}
-            for name, cmd in cmds.viewitems():
-                if cmd._indico_plugin:
-                    self._wrap_in_plugin_context(cmd._indico_plugin, cmd)
-                rv[name] = cmd
-        except Exception as exc:
-            if 'No indico config found' not in unicode(exc):
-                click.echo(click.style('Loading plugin commands failed:', fg='red', bold=True))
-                click.echo(click.style(traceback.format_exc(), fg='red'))
-            rv = {}
-        self._indico_plugin_commands = rv
-        return rv
-
-    def get_command(self, ctx, name):
-        rv = AppGroup.get_command(self, ctx, name)
-        if rv is not None:
-            return rv
-        return self._get_indico_plugin_commands(ctx).get(name)
-
-    def list_commands(self, ctx):
-        rv = set(click.Group.list_commands(self, ctx))
-        rv.update(self._get_indico_plugin_commands(ctx))
-        return sorted(rv)
-
-
-class LazyGroup(click.Group):
-    def __init__(self, import_name, **kwargs):
-        self._import_name = import_name
-        super(LazyGroup, self).__init__(**kwargs)
-
-    @cached_property
-    def _impl(self):
-        module, name = self._import_name.split(':', 1)
-        return getattr(import_module(module), name)
-
-    def get_command(self, ctx, cmd_name):
-        return self._impl.get_command(ctx, cmd_name)
-
-    def list_commands(self, ctx):
-        return self._impl.list_commands(ctx)
-
-    def invoke(self, ctx):
-        return self._impl.invoke(ctx)
-
-    def get_usage(self, ctx):
-        return self._impl.get_usage(ctx)
-
-    def get_params(self, ctx):
-        return self._impl.get_params(ctx)
 
 
 @click.group(cls=IndicoFlaskGroup)
@@ -128,6 +45,13 @@ def cli():
     command line.
     """
 
+
+# All core cli commands (including groups with subcommands) are
+# registered in this file and imported lazily - either using a
+# LazyGroup or by importing the actual logic of a command inside
+# the command's function here.  This allows running the cli even
+# if there's no indico config available or importing indico is
+# broken for other reasons.
 
 @cli.group(cls=LazyGroup, import_name='indico.cli.setup:cli')
 def setup():
